requirements.txt
streamlit>=1.32.0
yfinance>=0.2.37
nsepython>=2.9
pandas>=2.0.0
numpy>=1.26.0
pandas-ta>=0.3.14b
openpyxl>=3.1.2
APScheduler>=3.10.4
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.0
plotly>=5.18.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

